Refactor and separate all codebase
